+++ src/main/java/cart/domain/coupon/discountPolicy/DiscountPolicy.java
@@ -2,9 +2,11 @@ package cart.domain.coupon.discountPolicy;
 
 import cart.domain.Money;
 
+import java.math.BigDecimal;
+
 public interface DiscountPolicy {
 
-    Money discount(final Money money, final long value);
+    Money discount(final Money money, final BigDecimal value);
 
     String getName();
 }
