+++ src/main/java/cart/domain/coupon/discountPolicy/PricePolicy.java
@@ -2,10 +2,12 @@ package cart.domain.coupon.discountPolicy;
 
 import cart.domain.Money;
 
+import java.math.BigDecimal;
+
 public class PricePolicy implements DiscountPolicy {
 
     @Override
-    public Money discount(final Money orderPrice, final long value) {
+    public Money discount(final Money orderPrice, final BigDecimal value) {
         return new Money(value);
     }
