+++ src/main/java/cart/entity/OrderItemEntity.java
@@ -3,16 +3,18 @@ package cart.entity;
 import cart.domain.Money;
 import cart.domain.order.OrderItem;
 
+import java.math.BigDecimal;
+
 public class OrderItemEntity {
 
     private final Long id;
     private final String name;
     private final String imageUrl;
-    private final long price;
+    private final BigDecimal price;
     private final Integer quantity;
     private final Long orderId;
 
-    public OrderItemEntity(final Long id, final String name, final String imageUrl, final long price, final Integer quantity, final Long orderId) {
+    public OrderItemEntity(final Long id, final String name, final String imageUrl, final BigDecimal price, final Integer quantity, final Long orderId) {
         this.id = id;
         this.name = name;
         this.imageUrl = imageUrl;
@@ -21,7 +23,7 @@ public class OrderItemEntity {
         this.orderId = orderId;
     }
 
-    public OrderItemEntity(final String name, final String imageUrl, final long price, final Integer quantity, final Long orderId) {
+    public OrderItemEntity(final String name, final String imageUrl, final BigDecimal price, final Integer quantity, final Long orderId) {
         this(null, name, imageUrl, price, quantity, orderId);
     }
 
@@ -52,7 +54,7 @@ public class OrderItemEntity {
         return imageUrl;
     }
 
-    public long getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
