+++ src/test/java/cart/dao/CartItemDaoTest.java
@@ -11,6 +11,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -40,7 +41,7 @@ class CartItemDaoTest {
     void 장바구니에_상품을_추가한다() {
         // given
         final MemberEntity member = memberDao.insert(new MemberEntity("pizza1@pizza.com", "password1"));
-        final ProductEntity product = productDao.insert(new ProductEntity("치즈피자1", "1.jpg", 8900L));
+        final ProductEntity product = productDao.insert(new ProductEntity("치즈피자1", "1.jpg", BigDecimal.valueOf(8900L)));
         final CartItemEntity cartItemEntity = new CartItemEntity(member.getId(), product.getId(), 1);
 
         // when
@@ -55,9 +56,9 @@ class CartItemDaoTest {
         // given
         final MemberEntity member1 = memberDao.insert(new MemberEntity("pizza1@pizza.com", "password"));
         final MemberEntity member2 = memberDao.insert(new MemberEntity("pizza2@pizza.com", "password"));
-        final ProductEntity product1 = productDao.insert(new ProductEntity("치즈피자1", "1.jpg", 8900L));
-        final ProductEntity product2 = productDao.insert(new ProductEntity("치즈피자2", "2.jpg", 8900L));
-        final ProductEntity product3 = productDao.insert(new ProductEntity("치즈피자3", "3.jpg", 8900L));
+        final ProductEntity product1 = productDao.insert(new ProductEntity("치즈피자1", "1.jpg", BigDecimal.valueOf(8900L)));
+        final ProductEntity product2 = productDao.insert(new ProductEntity("치즈피자2", "2.jpg", BigDecimal.valueOf(8900L)));
+        final ProductEntity product3 = productDao.insert(new ProductEntity("치즈피자3", "3.jpg", BigDecimal.valueOf(8900L)));
 
         cartItemDao.insert(new CartItemEntity(member1.getId(), product1.getId(), 1));
         cartItemDao.insert(new CartItemEntity(member1.getId(), product2.getId(), 1));
@@ -74,7 +75,7 @@ class CartItemDaoTest {
     void 사용자_아이디와_삭제할_장바구니의_상품_아이디를_받아_장바구니_항목을_제거한다() {
         // given
         final MemberEntity member = memberDao.insert(new MemberEntity("pizza1@pizza.com", "password1"));
-        final ProductEntity product = productDao.insert(new ProductEntity("치즈피자1", "1.jpg", 8900L));
+        final ProductEntity product = productDao.insert(new ProductEntity("치즈피자1", "1.jpg", BigDecimal.valueOf(8900L)));
         final CartItemEntity cartItemEntity = new CartItemEntity(member.getId(), product.getId(), 1);
 
         // when
@@ -88,7 +89,7 @@ class CartItemDaoTest {
     void 장바구니의_상품의_수량을_변경한다() {
         // given
         final MemberEntity member = memberDao.insert(new MemberEntity("pizza1@pizza.com", "password1"));
-        final ProductEntity product = productDao.insert(new ProductEntity("치즈피자1", "1.jpg", 8900L));
+        final ProductEntity product = productDao.insert(new ProductEntity("치즈피자1", "1.jpg", BigDecimal.valueOf(8900L)));
         final CartItemEntity cartItemEntity = new CartItemEntity(member.getId(), product.getId(), 1);
         final CartItemEntity savedCartItemEntity = cartItemDao.insert(cartItemEntity);
